Feature/content statistics summary (#3932)
* Added State counts for entities and DARs, modified DTO to have metric properties

* Implemented UI components for Content stats

* Added more statistics item UI, renamed content-stats -> entity-statistics-summary

* Added requireIndexing in DS state + related CRUD ops

* Added IndexHealth resource to return docs needing indexing, DS modified to index from a list of IDs.

* Added indexing by IDs for Networks and datasets

* Added indexing bu IDs for Study, Project and Network

* Added more HealIndex, Popup for indxing entities and more messages.

* Chanegd strategy of identifying entities require indexing due to performance issues.

* Corrected import

* Added warning messages, enahnced DAR counts, added config options to hide some stats.

* Added Fr translations and some message reformatting.

* Code cleanups

* Used a regular alert so unpublishing a DS removes the warning.

+++ mica-core/src/main/java/org/obiba/mica/web/model/DataAccessRequestDtos.java
@@ -327,8 +327,8 @@ class DataAccessRequestDtos {
       builder.setAmendmentsSummary(
         Mica.DataAccessRequestDto.AmendmentsSummaryDto.newBuilder()
           .setId(map.get("_id") + "")
-          .setPending((double) map.get("pending"))
-          .setTotal((double) map.get("total"))
+          .setPending((int)map.get("pending"))
+          .setTotal((int)map.get("total"))
           .setLastModifiedDate(map.get("lastModified") instanceof Date ? ((Date) map.get("lastModified")).toInstant().atOffset(ZoneOffset.systemDefault().getRules().getOffset(Instant.now())).toString() : "")
           .build()
       );
